+++ src/settings.js
@@ -488,6 +488,13 @@ const SETTINGS = {
           type: Boolean,
           default: false,
         },
+        "append-chris-premade-effect-description": {
+          name: "ddb-importer.settings.append-chris-premade-effect-description.name",
+          hint: "ddb-importer.settings.append-chris-premade-effect-description.hint",
+          config: true,
+          type: Boolean,
+          default: false,
+        }
       },
       // ????
       MISC: {