+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Monster Effects: Beholder Legendary Eye Ray action will now roll Eye Rays macro.
 - Thunder Gauntlets and Lightning Launcher change on DDB meant they are no longer importing as weapons on newer characters.
 - Slight tweak for ability bonus effects (such as belt of dwarfenkind) if not using DAE.
+- Global setting to append the  Chris Premades description to the end of updated features etc.
+- Sacred Weapon Macro fix.
 
 # 5.1.24